api/pyproject.toml
[tool.poetry]
name = "nekos api"
version = "2.0.0-Alpha"
description = "An open-source, free and public anime images REST/GraphQL API."
authors = ["Neki <nekidev@nekosapi.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
strawberry-graphql = "^0.155.2"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
djangorestframework-jsonapi = "^6.0.0"
django-filter = "^22.1"
django-admin-interface = "^0.24.2"
django-oauth-toolkit = "^2.2.0"
django-hosts = "^5.2"
django-ratelimit = "^4.0.0"
django-resized = "^1.0.2"
django-dynamic-filenames = "^1.3.1"
django-cleanup = "^6.0.0"
redis = "^4.4.2"
django-bunny-storage = "^0.1.2"
Pillow = "^9.4.0"
requests-oauthlib = "^1.3.1"
channels = "^4.0.0"
daphne = "^4.0.0"
colorthief = "^0.2.1"
django-logentry-admin = "^1.1.0"
channels-redis = "^4.1.0"
schema = "^0.7.5"
validators = "^0.20.0"
tldextract = "^3.4.0"
dnspython = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
